Fix browsers freezing on re-calculating styles
Previously each change lead to a new "style" node being created and
appended to the DOM. Old style nodes were never reused. The list
quickly grows to a few hundred style nodes when dragging a few settings.
Re-calculating hundreds of stylesheets is expensive and can lead to
browsers freezing for a few seconds.

Therefore the fix is to always re-use the same style node and mark the
DOM node creation as a layout effect for React to schedule accordingly.